Revert "Fixes if size_t not uint64_t"
This reverts commit cf95cefa03d1d26d597f106afb6c15bdfa9cb277.

partitionRange works now.

+++ main/src/init/grid.hpp
@@ -53,14 +53,14 @@ inline auto partitionRange(size_t R, size_t i, size_t N)
     size_t r = R % N;
     if (i < r)
     {
-        uint64_t start = (s + 1) * i;
-        uint64_t end   = start + s + 1;
+        size_t start = (s + 1) * i;
+        size_t end   = start + s + 1;
         return std::make_tuple(start, end);
     }
     else
     {
-        uint64_t start = (s + 1) * r + s * (i - r);
-        uint64_t end   = start + s;
+        size_t start = (s + 1) * r + s * (i - r);
+        size_t end   = start + s;
         return std::make_tuple(start, end);
     }
 }
